# api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for task tracker",
        "title": "Task Tracker API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/api/project": {
            "post": {
                "description": "Creates a new project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Projects"
                ],
                "summary": "Create new projects",
                "parameters": [
                    {
                        "description": "Project data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ProjectCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created project ID",
                        "schema": {
                            "$ref": "#/definitions/model.ProjectCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/project/list": {
            "get": {
                "description": "Return all projects in system",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Projects"
                ],
                "summary": "Get all projects",
                "responses": {
                    "200": {
                        "description": "Tasks",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.ProjectListResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.ResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apiresponse.ResponseError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "model.ProjectCreateRequest": {
            "type": "object",
            "properties": {
                "member_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "60601fee-2bf1-4721-ae6f-7636e79a0cba"
                    ]
                },
                "name": {
                    "type": "string",
                    "example": "Project name"
                },
                "viewer_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "60601fee-2bf1-4721-ae6f-7636e79a0cba"
                    ]
                }
            }
        },
        "model.ProjectCreateResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "60601fee-2bf1-4721-ae6f-7636e79a0cba"
                }
            }
        },
...
        "model.ProjectListResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "OAuth2Keycloak": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "http://localhost:8155/realms/fmgo/protocol/openid-connect/token"
        }
    },
    "security": [
        {
            "OAuth2Keycloak": []
        }
    ]
}